changed ideal to float

+++ src/Components/stair.js
@@ -11,8 +11,8 @@ const Stair = (props) => {
   let totalRun = parseInt(props.totalRunft) + parseInt(props.totalRunin) + parseFloat(props.totalRunfr);
   let totalRise = parseInt(props.totalRiseft) + parseInt(props.totalRisein) + parseFloat(props.totalRisefr);
 
-  let idealRun = parseInt(props.idealRunin) + parseFloat(props.idealRunfr);
-  let idealRise = parseInt(props.idealRisein) + parseFloat(props.idealRisefr);
+  let idealRun = parseFloat(props.idealRunin) + parseFloat(props.idealRunfr);
+  let idealRise = parseFloat(props.idealRisein) + parseFloat(props.idealRisefr);
 
   let xCount = parseInt(totalRun/idealRun);
   let yCount = parseInt(totalRise/idealRise);
